moved RelatedPokemons into PokemoDetail folder
