+++ src/components/PokemonDetails/RelatedPokemon.vue
